refactor: reorganize source into src, scripts, and test
Give each directory a single role: src/ for published package source,
scripts/ for build tooling, and test/ for the test harness.

- Split color-p3.ts into a src/color/ module (srgb, p3, and a barrel).
- Split palette.ts into src/palettes.ts and per-variant src/roles/.
- Rename theme.ts/zed-theme.ts (and their makeTheme/makeZedThemeFamily
  exports) to createTheme.ts/createZedTheme.ts.
- Move the preview builder to src/previews/ and add a scripts/
  createPreviews.ts generator that renders previews to disk.
- Move build.ts to scripts/, and the VSIX packager to scripts/
  buildVsCodePackage.ts with its marketplace README colocated.
- Move the test harness to test/.

Update package.json script paths, broaden the tsconfig include to
src+scripts+test (dropping rootDir), and fix the publish workflow's
README path. Behavior-preserving: the six themes regenerate identically.
