+++ .github/workflows/publish.yml
@@ -29,7 +29,7 @@ jobs:
         run: jq '.name = "pierre-theme"' package.json > tmp.json && mv tmp.json package.json
 
       - name: Swap in VSCE README
-        run: cp README.package.md README.md
+        run: cp scripts/README.package.md README.md
 
       - name: Publish to VS Marketplace
         env:
