+++ src/roles/dark.ts
@@ -1,0 +1,79 @@
+import type { Roles } from "./Roles";
+import { palettes } from "../palettes";
+const { gray, neutral, red, vermillion, orange, amber, yellow, lime, green, jade, mint, teal, cyan, blue, indigo, violet, purple, magenta, pink, rose, brown } = palettes;
+
+export const dark: Roles = {
+  bg: {
+    editor: neutral["1040"],
+    window: neutral["1000"],
+    inset: neutral["980"],
+    elevated: neutral["1020"]
+  },
+  fg: {
+    base: neutral["020"],
+    fg1: neutral["200"],
+    fg2: neutral["400"],
+    fg3: neutral["600"],
+    fg4: neutral["700"]
+  },
+  border: {
+    window: neutral["1040"],
+    editor: neutral["980"],
+    indentGuide: neutral["980"],
+    indentGuideActive: neutral["960"],
+    inset: neutral["980"],
+    elevated: neutral["980"]
+  },
+  accent: {
+    primary: blue["500"],
+    link: blue["500"],
+    subtle: blue["950"],
+    contrastOnAccent: neutral["1040"]
+  },
+  states: {
+    merge: indigo["500"],
+    success: jade["500"],
+    danger: red["500"],
+    warn: yellow["500"],
+    info: cyan["500"]
+  },
+  syntax: {
+    comment: neutral["600"],
+    string: green["400"],
+    number: cyan["400"],
+    keyword: pink["400"],
+    regexp: teal["400"],
+    func: indigo["400"],
+    type: purple["400"],
+    variable: orange["400"],
+    // Extended token types
+    operator: cyan["500"],
+    punctuation: neutral["700"],
+    constant: yellow["400"],
+    parameter: neutral["400"],
+    namespace: amber["500"],
+    decorator: blue["400"],
+    escape: mint["400"],
+    invalid: neutral["020"],
+    tag: vermillion["400"],
+    attribute: jade["400"]
+  },
+  ansi: {
+    black: neutral["1000"],
+    red: red["500"],
+    green: green["500"],
+    yellow: yellow["500"],
+    blue: blue["500"],
+    magenta: magenta["500"],
+    cyan: cyan["500"],
+    white: neutral["300"],
+    brightBlack: neutral["1000"],
+    brightRed: red["500"],
+    brightGreen: lime["500"],
+    brightYellow: yellow["500"],
+    brightBlue: blue["500"],
+    brightMagenta: magenta["500"],
+    brightCyan: cyan["500"],
+    brightWhite: neutral["300"]
+  }
+};
