+++ src/roles/darkSoft.ts
@@ -1,0 +1,79 @@
+import type { Roles } from "./Roles";
+import { palettes } from "../palettes";
+const { gray, neutral, red, vermillion, orange, amber, yellow, lime, green, jade, mint, teal, cyan, blue, indigo, violet, purple, magenta, pink, rose, brown } = palettes;
+
+export const darkSoft: Roles = {
+  bg: {
+    editor: neutral["1000"],
+    window: neutral["1020"],
+    inset: neutral["960"],
+    elevated: neutral["980"]
+  },
+  fg: {
+    base: neutral["200"],
+    fg1: neutral["300"],
+    fg2: neutral["500"],
+    fg3: neutral["700"],
+    fg4: neutral["800"]
+  },
+  border: {
+    window: neutral["980"],
+    editor: neutral["940"],
+    indentGuide: neutral["960"],
+    indentGuideActive: neutral["940"],
+    inset: neutral["940"],
+    elevated: neutral["960"]
+  },
+  accent: {
+    primary: blue["400"],
+    link: blue["400"],
+    subtle: blue["900"],
+    contrastOnAccent: neutral["1000"]
+  },
+  states: {
+    merge: indigo["400"],
+    success: jade["400"],
+    danger: red["400"],
+    warn: yellow["400"],
+    info: cyan["400"]
+  },
+  syntax: {
+    comment: neutral["700"],
+    string: green["300"],
+    number: cyan["300"],
+    keyword: pink["300"],
+    regexp: teal["300"],
+    func: indigo["300"],
+    type: purple["300"],
+    variable: orange["300"],
+    // Extended token types
+    operator: cyan["400"],
+    punctuation: neutral["600"],
+    constant: yellow["300"],
+    parameter: neutral["500"],
+    namespace: amber["400"],
+    decorator: blue["300"],
+    escape: mint["300"],
+    invalid: neutral["200"],
+    tag: vermillion["300"],
+    attribute: jade["300"]
+  },
+  ansi: {
+    black: neutral["1000"],
+    red: red["500"],
+    green: green["500"],
+    yellow: yellow["500"],
+    blue: blue["500"],
+    magenta: magenta["500"],
+    cyan: cyan["500"],
+    white: neutral["300"],
+    brightBlack: neutral["1000"],
+    brightRed: red["500"],
+    brightGreen: lime["500"],
+    brightYellow: yellow["500"],
+    brightBlue: blue["500"],
+    brightMagenta: magenta["500"],
+    brightCyan: cyan["500"],
+    brightWhite: neutral["300"]
+  }
+};
