+++ src/roles/light.ts
@@ -1,0 +1,80 @@
+import type { Roles } from "./Roles";
+import { palettes } from "../palettes";
+const { gray, neutral, red, vermillion, orange, amber, yellow, lime, green, jade, mint, teal, cyan, blue, indigo, violet, purple, magenta, pink, rose, brown } = palettes;
+
+export const light: Roles = {
+  bg: {
+    editor: "#ffffff",
+    window: neutral["060"],
+    inset: neutral["080"],
+    elevated: neutral["040"]
+  },
+  fg: {
+    base: neutral["1040"],
+    fg1: neutral["900"],
+    fg2: neutral["800"],
+    fg3: neutral["600"],
+    fg4: neutral["500"]
+  },
+  border: {
+    window: neutral["100"],
+    editor: neutral["200"],
+    indentGuide: neutral["100"],
+    indentGuideActive: neutral["200"],
+    inset: neutral["200"],
+    elevated: neutral["100"]
+  },
+  accent: {
+    primary: blue["500"],
+    link: blue["500"],
+    subtle: blue["100"],
+    contrastOnAccent: "#ffffff"
+  },
+  states: {
+    merge: indigo["600"],
+    success: jade["600"],
+    danger: red["600"],
+    warn: yellow["600"],
+    info: cyan["600"]
+  },
+  syntax: {
+    comment: neutral["600"],
+    string: green["600"],
+    number: cyan["600"],
+    keyword: pink["600"],
+    regexp: teal["600"],
+    func: indigo["600"],
+    type: purple["600"],
+    variable: orange["600"],
+    // Extended token types
+    operator: cyan["500"],
+    punctuation: neutral["700"],
+    constant: yellow["600"],
+    parameter: neutral["700"],
+    namespace: amber["600"],
+    decorator: blue["600"],
+    escape: mint["600"],
+    invalid: neutral["1040"],
+    tag: vermillion["600"],
+    attribute: jade["600"]
+  },
+  ansi: {
+    black: neutral["980"],
+    red: red["600"],
+    green: jade["600"],
+    yellow: yellow["600"],
+    blue: blue["600"],
+    magenta: magenta["600"],
+    cyan: cyan["600"],
+    white: neutral["300"],
+    // make bright colors match the non-bright counterparts
+    brightBlack: neutral["980"],
+    brightRed: red["600"],
+    brightGreen: lime["600"],
+    brightYellow: yellow["600"],
+    brightBlue: blue["600"],
+    brightMagenta: magenta["600"],
+    brightCyan: cyan["600"],
+    brightWhite: neutral["300"]
+  }
+};
