+++ src/roles/lightSoft.ts
@@ -1,0 +1,79 @@
+import type { Roles } from "./Roles";
+import { palettes } from "../palettes";
+const { gray, neutral, red, vermillion, orange, amber, yellow, lime, green, jade, mint, teal, cyan, blue, indigo, violet, purple, magenta, pink, rose, brown } = palettes;
+
+export const lightSoft: Roles = {
+  bg: {
+    editor: "#ffffff",
+    window: neutral["040"],
+    inset: neutral["060"],
+    elevated: neutral["020"]
+  },
+  fg: {
+    base: neutral["800"],
+    fg1: neutral["700"],
+    fg2: neutral["600"],
+    fg3: neutral["500"],
+    fg4: neutral["400"]
+  },
+  border: {
+    window: neutral["080"],
+    editor: neutral["100"],
+    indentGuide: neutral["080"],
+    indentGuideActive: neutral["100"],
+    inset: neutral["200"],
+    elevated: neutral["100"]
+  },
+  accent: {
+    primary: blue["500"],
+    link: blue["500"],
+    subtle: blue["100"],
+    contrastOnAccent: "#ffffff"
+  },
+  states: {
+    merge: indigo["500"],
+    success: jade["500"],
+    danger: red["500"],
+    warn: yellow["500"],
+    info: cyan["500"]
+  },
+  syntax: {
+    comment: neutral["500"],
+    string: green["500"],
+    number: cyan["500"],
+    keyword: pink["400"],
+    regexp: teal["500"],
+    func: indigo["400"],
+    type: purple["400"],
+    variable: orange["500"],
+    // Extended token types
+    operator: cyan["400"],
+    punctuation: neutral["600"],
+    constant: yellow["500"],
+    parameter: neutral["600"],
+    namespace: amber["500"],
+    decorator: blue["400"],
+    escape: mint["500"],
+    invalid: neutral["1000"],
+    tag: vermillion["500"],
+    attribute: jade["500"]
+  },
+  ansi: {
+    black: neutral["980"],
+    red: red["500"],
+    green: green["500"],
+    yellow: yellow["500"],
+    blue: blue["500"],
+    magenta: magenta["500"],
+    cyan: cyan["500"],
+    white: neutral["300"],
+    brightBlack: neutral["980"],
+    brightRed: red["500"],
+    brightGreen: lime["500"],
+    brightYellow: yellow["500"],
+    brightBlue: blue["500"],
+    brightMagenta: magenta["500"],
+    brightCyan: cyan["500"],
+    brightWhite: neutral["300"]
+  }
+};
